Refactor make_request and parsers into own files
