+++ pySpaceTraders/utils/parsers.py
@@ -1,45 +1,4 @@
-import logging
-from typing import Optional
-
-import requests
-from ratelimit import limits, sleep_and_retry
-
-from pySpaceTraders.constants import __version__, V2_STARTRADERS_URL, REQUEST_TYPES
-from pySpaceTraders.models import agent, cargo, contract, enums, errors, factions, status
-
-
-@sleep_and_retry
-@limits(calls=2, period=1.2)
-def make_request(
-        method: str, endpoint: str, token: Optional[str] = "", params: Optional[dict] = ""
-):
-    """
-    ### Parameters
-    - method: str
-        - Request Method (GET, POST, PUT, DELETE)
-    - endpoint: str
-        - Endpoint you are trying to reach
-    - *token: Optional[str] (Defaults Blank)
-        - Your JWT Token, only needed for authenticated endpoints.
-    - *params: Optional[dict] (Defaults Blank)
-        - Any payload data.
-    """
-    headers = {"User-Agent": f"pySpaceTraders/{__version__}"}
-    if token:
-        headers["Authorization"] = f"Bearer {token}"
-    method = method.upper()
-
-    if method not in REQUEST_TYPES:
-        logging.exception(f"Invalid request method: {method}")
-        return None
-
-    endpoint = V2_STARTRADERS_URL + endpoint
-    if method == "GET":
-        return requests.get(endpoint, headers=headers, params=params)
-    elif method == "POST":
-        return requests.post(endpoint, headers=headers, data=params)
-    elif method == "PATCH":
-        return requests.patch(endpoint, headers=headers, data=params)
+from pySpaceTraders.models import cargo, contract, errors, factions, status
 
 
 def parse_error(response):
@@ -59,7 +18,7 @@ def parse_contract(contract_in: dict) -> contract.Contract:
     return contract.Contract(**contract_in)
 
 
-def parse_cargo(cargo_in: dict) -> contract.Contract:
+def parse_cargo(cargo_in: dict) -> cargo.Cargo:
     cargo_in["inventory"] = [cargo.Item(**item) for item in cargo_in["inventory"]]
     data = cargo.Cargo(**cargo_in)
     return data
